requirements.txt
streamlit
huggingface_hub
llama-cpp-python
sentence-transformers
transformers
langchain
requests
pydantic